Fix #2
Now the generator looks to see if /lib/impact exists in the source folder and uses that path instead. Otherwise the source folder is used. This should get the license key/download working, and also allow the source folder to be set to a root folder of a game that already contains the impact engine files.